requirements.txt
flask
transformers
accelerate
bitsandbytes
peft
sentencepiece
pyyaml
datasets
tqdm
beautifulsoup4
requests
lxml
spacy
